backend/pyproject.toml
[project]
name = "maxirest-backend"
version = "0.1.0"
description = "Maxirest MVP backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn[standard]==0.35.0",
  "sqlalchemy==2.0.43",
  "psycopg[binary]==3.2.9",
  "pydantic-settings==2.10.1",
  "python-multipart==0.0.20",
  "PyJWT==2.10.1",
  "email-validator==2.2.0",
  "minio==7.2.16",
  "pytest==8.4.1",
  "httpx==0.28.1",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = [".", "app"]
testpaths = ["app/tests"]
